logging

+++ lib/cloghandler/cloghandler.py
@@ -1,30 +1,31 @@
 import os
-from logging import Handler
+import Queue
+from logging import Handler, StreamHandler
 from logging.handlers import TimedRotatingFileHandler
-
 import time
-
 from portalocker import lock, unlock, LOCK_EX, LOCK_NB, LockException
+from static import log_path
 
 FORCE_ABSOLUTE_PATH = False
+
+# ---------------------------------------------------------------------------
+#   Miscellaneous module data
+# ---------------------------------------------------------------------------
+try:
+    unicode
+    _unicode = True
+except NameError:
+    _unicode = False
 
 
 class ConcurrentTimedRotatingFileHandler(TimedRotatingFileHandler):
     def __init__(self, filename, when='h', interval=1, backupCount=0,
-                 encoding=None, delay=0, utc=0, debug=False, supress_abs_warn=False):
+                 encoding=None, delay=0, utc=0, debug=False, supress_abs_warn=True):
         # if the given filename contains no path, we make an absolute path
-        if not os.path.isabs(filename):
-            if FORCE_ABSOLUTE_PATH or \
-                    not os.path.split(filename)[0]:
-                filename = os.path.abspath(filename)
-            elif not supress_abs_warn:
-                from warnings import warn
-
-                warn("The given 'filename' should be an absolute path.  If your "
-                     "application calls os.chdir(), your logs may get messed up. "
-                     "Use 'supress_abs_warn=True' to hide this message.")
+        filename = os.path.join(log_path, filename)
         TimedRotatingFileHandler.__init__(self, filename, when, interval, backupCount, encoding, delay, utc)
         self.currentSuffix = time.strftime(self.suffix, time.localtime())
+        self.doRollover()
         self.stream_lock = open(filename.replace(".log", "") + ".lock", "w")
 
     def acquire(self):
@@ -39,9 +40,8 @@ class ConcurrentTimedRotatingFileHandler(TimedRotatingFileHandler):
 
     def release(self):
         try:
-            self.stream.flush()
-        except:
-            pass
+            if self.stream:
+                self.stream.flush()
         finally:
             try:
                 unlock(self.stream_lock)
@@ -62,8 +62,11 @@ class ConcurrentTimedRotatingFileHandler(TimedRotatingFileHandler):
         dfn = self.baseFilename + "." + self.currentSuffix
         # dst file exists infers that the other process has renamed the file
         # so just open the basefile again
-        if not os.path.exists(dfn):
-            os.rename(self.baseFilename, dfn)
+        if not os.path.exists(dfn) and os.path.exists(self.baseFilename):
+            try:
+                os.rename(self.baseFilename, dfn)
+            except Exception:
+                pass
         self.mode = 'a'
         self.stream = self._open()
         self.currentSuffix = time.strftime(self.suffix, time.localtime())
